Add SPV wallet kit project, move NodeWithWalletTest
